combine devices + files
