+++ src/devices/bus/spectrum/exp.cpp
@@ -156,17 +156,16 @@ void spectrum_expansion_slot_device::mreq_w(offs_t offset, uint8_t data)
 // slot devices
 #include "beta.h"
 #include "beta128.h"
-#include "disciple.h"
 #include "intf1.h"
 #include "intf2.h"
 #include "fuller.h"
 #include "kempjoy.h"
 #include "melodik.h"
 #include "mface.h"
+#include "mgt.h"
 #include "mikroplus.h"
 #include "opus.h"
 #include "plus2test.h"
-#include "plusd.h"
 #include "protek.h"
 #include "specdrum.h"
 #include "uslot.h"
